server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
nltk==3.6.2
pydantic==1.8.2
fastapi==0.68.0
streamlit==1.10.0
matplotlib==3.5.1
seaborn==0.11.2
python-dotenv==0.19.1
requests==2.26.0
openai>=1.3.0

# After installing requirements, run this to set up NLTK data:
# python setup_nltk.py